examples/aip-agent-quickstart/pyproject.toml
[tool.poetry]
name = "aip-agent-quickstart"
version = "0.1.0"
description = "AIP Agent Quickstart Example"
authors = ["Raymond Christopher <raymond.christopher@gdplabs.id>", "Fachriza Adhiatma <fachriza.d.adhiatma@gdplabs.id>", "Christian Trisno Sen Long Chen <christian.t.s.l.chen@gdplabs.id>"]
readme = "README.md"
packages = [{include = "aip_agent_quickstart"}]

[tool.poetry.dependencies]
bosa-connectors-binary = "^0.1.0"
python = ">=3.13,<3.14"
gllm-agents-binary = "^0.2.13"
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langgraph = "^0.2.16"
nest-asyncio = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
ruff = "^0.6.7"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = ["B026", "C400", "C901", "E203", "E266", "F403", "F821"]
select = ["B", "B9", "C", "D", "E", "F", "I", "PL", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "PLR"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[[tool.poetry.source]]
name = "gen-ai"
url = "https://asia-southeast2-python.pkg.dev/gdp-labs/gen-ai/simple/"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
